:wrench: Fix: Correct links to package web page
Use an emoji, if desired, to help visually identify the commit subject
Use an issue an commit type to assist with subject identification.
Summarize the the contents of the commit
Limit content to 1 line containing 50 characters

MAIN TEXT
What does the commit address?
- Be brief and descriptive
 - =<80 characters per line
 - One commit per issue
  - If overlong split changes across multiple commits
 - Use imperative
- Answer each question where needed
- Bullet points help reading but paragraphs are also acceptable.
- Links
 - Link / links to an related issue (if exists)
 - Link to pull request (if exists)
 - Link to fork (if exists)

How does it address the issue?
  - (if applicable)

What side effects does this change have?
  - (if applicable)